+++ CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Set leader to ` `. It is the space key.
+
+- `<LEADER>sv` - Split window veritcally
+- `<LEADER>sh` - Split window horizontally
+- `<LEADER>se` - Make splits to be of equal size
+- `<LEADER>sx` - Close current split
+
+- `<LEADER>to` - Open new tab
+- `<LEADER>tx` - Close current tab
+- `<LEADER>tn` - Go to next tab
+- `<LEADER>tp` - Go to previous tab
+
+- `Shift+H` - Decrease split horizontal size
+- `Shift+L` - Increase split horizontal size
+- `Shift+J` - Increase split vertical size
+- `Shift+K` - Decrease split vertical size
+
+- `<LEADER>nh` - Clear search highlights
+
+### Tests
+
+- Add test container for Fedora 41
+- Add test container for Ubuntu 24.10
+
 ## [0.1.0] - 2024-11-07
 
 ### Added